requirements.txt
# Tavily client
# (license depends on upstream; if this ever blocks you, we can call their REST API via requests instead)
tavily-python==0.5.4

# Groq official client
groq==0.9.0

# Groq client currently breaks with httpx>=0.28 (proxies kwarg)
httpx==0.27.2

# Utils
pydantic==2.7.4
requests==2.32.3
